Created a dynamic way to keep track of the student info
I created a folder name php to place any future php files we create.

I duplicated the homepage as a PHP file to add the PHP loop in it. Run
this file using MAMP on your computer

The "studentinfo.csv" file is where we are going to keep all the
infomation we get for the site. Edit in Excel and save. Refresh the
page to see the change.